+++ Test/ServerTelemetryChannel.Test/Shared.Tests/Implementation/ThrottlingTransmissionPolicyTest.cs
@@ -4,16 +4,11 @@
     using System.Diagnostics.Tracing;
     using System.Globalization;
     using System.Linq;
-    using System.Net;
-#if NETCOREAPP1_1
-    using System.Reflection;
-#endif
     using System.Threading;
+    using Microsoft.ApplicationInsights.Extensibility.Implementation;
     using Microsoft.ApplicationInsights.TestFramework;
     using Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.Helpers;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
-     
-    using TaskEx = System.Threading.Tasks.Task;    
 
     public class ThrottlingTransmissionPolicyTest
     {
@@ -51,8 +46,10 @@
 
                 transmitter.OnTransmissionSent(
                     new TransmissionProcessedEventArgs(
-                        new StubTransmission(),
-                    CreateThrottledResponse(ResponseCodePaymentRequired, 1)));
+                        new StubTransmission(), null, new HttpWebResponseWrapper()
+                        {
+                            StatusCode = ResponseCodePaymentRequired
+                        }));
 
                 Assert.IsNull(policy.MaxSenderCapacity);
                 Assert.IsNull(policy.MaxBufferCapacity);
@@ -60,7 +57,7 @@
             }
 
             [TestMethod]
-            public void AssertUnsupportedResponseCodeDoesntChangeCapacity()
+            public void AssertUnsupportedResponseCodeDoesnotChangeCapacity()
             {
                 var transmitter = new StubTransmitter();
                 transmitter.OnApplyPolicies = () =>
@@ -73,8 +70,10 @@
 
                 transmitter.OnTransmissionSent(
                     new TransmissionProcessedEventArgs(
-                        new StubTransmission(),
-                    CreateThrottledResponse(ResponseCodeUnsupported, 1)));
+                        new StubTransmission(), null, new HttpWebResponseWrapper()
+                        {
+                            StatusCode = ResponseCodeUnsupported
+                        }));
 
                 Assert.IsNull(policy.MaxSenderCapacity);
                 Assert.IsNull(policy.MaxBufferCapacity);
@@ -84,7 +83,7 @@
             [TestMethod]
             public void CannotParseRetryAfterWritesToEventSource()
             {
-                const string UnparsableDate = "no one can parse me! :)";
+                const string unparsableDate = "no one can parse me! :)";
 
                 var transmitter = new StubTransmitter();
                 var policy = new ThrottlingTransmissionPolicy();
@@ -97,41 +96,15 @@
 
                     transmitter.OnTransmissionSent(
                         new TransmissionProcessedEventArgs(
-                            new StubTransmission(),
-                        CreateThrottledResponse(ResponseCodeTooManyRequests, UnparsableDate)));
-
+                            new StubTransmission(),null, new HttpWebResponseWrapper()
+                            {
+                                StatusCode = ResponseStatusCodes.ResponseCodeTooManyRequests,
+                                RetryAfterHeader = unparsableDate
+                            })
+                        );
                     EventWrittenEventArgs trace = listener.Messages.First(args => args.EventId == 24);
-                    Assert.AreEqual(UnparsableDate, (string)trace.Payload[0]);
+                    Assert.AreEqual(unparsableDate, (string)trace.Payload[0]);
                 }
-            }
-
-            private static WebException CreateThrottledResponse(int throttledStatusCode, int retryAfter)
-            {
-                return CreateThrottledResponse(throttledStatusCode, retryAfter.ToString(CultureInfo.InvariantCulture));
-            }
-
-            private static WebException CreateThrottledResponse(int throttledStatusCode, string retryAfter)
-            {
-                var responseHeaders = new WebHeaderCollection();
-                responseHeaders[HttpResponseHeader.RetryAfter] = retryAfter;
-
-#if NETCOREAPP1_1
-                System.Net.Http.HttpResponseMessage responseMessage = new System.Net.Http.HttpResponseMessage((HttpStatusCode)throttledStatusCode);
-                
-                ConstructorInfo ctor = typeof(HttpWebResponse).GetConstructors(BindingFlags.Instance | BindingFlags.NonPublic)[0];
-                HttpWebResponse webResponse = (HttpWebResponse)ctor.Invoke(new object[] { responseMessage, null, null });
-
-                typeof(HttpWebResponse).GetField("_webHeaderCollection", BindingFlags.NonPublic | BindingFlags.Instance).SetValue(webResponse, (WebHeaderCollection)responseHeaders);
-
-                return new WebException("Transmitter Error", null, WebExceptionStatus.UnknownError, webResponse);                
-#else
-                var mockWebResponse = new Moq.Mock<HttpWebResponse>();
-                
-                mockWebResponse.SetupGet<HttpStatusCode>((webRes) => webRes.StatusCode).Returns((HttpStatusCode)throttledStatusCode);
-                mockWebResponse.SetupGet<WebHeaderCollection>((webRes) => webRes.Headers).Returns((WebHeaderCollection)responseHeaders);
-
-                return new WebException("Transmitter Error", null, WebExceptionStatus.UnknownError, mockWebResponse.Object);
-#endif
             }
 
             private void PositiveTest(int responseCode, int? expectedSenderCapacity, int? expectedBufferCapacity, int? expectedStorageCapacity)
@@ -156,8 +129,11 @@
 
                 transmitter.OnTransmissionSent(
                     new TransmissionProcessedEventArgs(
-                        new StubTransmission(),
-                    CreateThrottledResponse(responseCode, retryAfter)));
+                        new StubTransmission(), null, new HttpWebResponseWrapper()
+                        {
+                            StatusCode = responseCode,
+                            RetryAfterHeader = retryAfter
+                        }));
 
                 Assert.IsTrue(policyApplied.WaitOne(WaitForTheFirstApplyAsync));